Remove useless functions
