+++ react/next/components/SlideList.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 
-import { getIfSlideIsVisbile } from '../utils/index'
 import Slide from './Slide'
 
 interface Props {
@@ -9,6 +8,17 @@ interface Props {
   slidesToShow: number
   children: any
   itemClass?: string
+}
+
+/**
+ * Returns if the slide is visible or not
+ */
+const getIfSlideIsVisbile = (
+  index: number,
+  currentSlide: number,
+  slidesToShow: number
+): boolean => {
+  return index >= currentSlide && index < currentSlide + slidesToShow
 }
 
 /** List of all slides */
