+++ react/next/components/SliderTrack.tsx
@@ -4,7 +4,6 @@ import { transitionType } from '../typings/global'
 interface Props {
   className?: string
   transform: number
-  shouldRenderOnSSR: boolean
   transition: transitionType
 }
 
@@ -13,7 +12,6 @@ const SliderTrack: FC<Props> = ({
   className,
   transform,
   transition,
-  shouldRenderOnSSR,
 }) => (
   <div
     className={`${className} flex relative pa0 ma0`}
@@ -22,7 +20,6 @@ const SliderTrack: FC<Props> = ({
       transition: `transform ${transition.speed}ms ${transition.timing}`,
       transitionDelay: `${transition.delay}ms`,
       transform: `translate3d(${transform}px, 0, 0)`,
-      overflow: shouldRenderOnSSR ? 'hidden' : 'unset',
     }}
   >
     {children}
