+++ react/next/utils/index.ts
@@ -1,13 +1,10 @@
 import { getWidthFromDeviceType, getItemClientSideWidth } from './elementWidth'
-import { getInitialState, getIfSlideIsVisbile } from './common'
 import { populateNextSlides } from './next'
 import { populatePreviousSlides } from './previous'
 
 export {
   getWidthFromDeviceType,
   getItemClientSideWidth,
-  getInitialState,
-  getIfSlideIsVisbile,
   populateNextSlides,
   populatePreviousSlides,
 }
